accessible/tests/mochitest/tree/test_gencontent.html

<!DOCTYPE html>
<html>

<head>
  <title>Generated content tests</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <style>
    .gentext:before {
      content: "START"
    }
    .gentext:after {
      content: "END"
    }
  </style>

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>

  <script type="application/javascript">
    function doTest() {
      // :before and :after pseudo styles
      var accTree = {
        role: ROLE_SECTION,
        children: [
          {
            role: ROLE_STATICTEXT,
            name: "START",
          },
          {
            role: ROLE_TEXT_LEAF,
            name: "MIDDLE",
          },
          {
            role: ROLE_STATICTEXT,
            name: "END",
          },
        ],
      };

      testAccessibleTree("gentext", accTree);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="Clean up our tree walker"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=530081">
    Mozilla Bug 530081
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <div class="gentext" id="gentext">MIDDLE</div>
</body>
</html>
